requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
yfinance==0.2.28
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
python-dotenv==1.0.0
groq==0.4.2
google-generativeai==0.3.2
beautifulsoup4==4.12.2
duckduckgo-search==3.9.6
plotly==5.17.0
kaleido==0.2.1
ta==0.11.0
uuid==1.30